+++ build_runner/test/generate/create_merged_dir_test.dart
@@ -35,12 +35,14 @@ main() {
       makeAssetId('a|lib/a.txt'): 'a',
       makeAssetId('a|web/b.txt'): 'b',
       makeAssetId('b|lib/c.txt'): 'c',
+      makeAssetId('a|foo/d.txt'): 'd',
     };
     final packageGraph = buildPackageGraph({
       rootPackage('a'): ['b'],
       package('b'): []
     });
     Directory tmpDir;
+    Directory anotherTmpDir;
     TestBuildEnvironment environment;
     InMemoryRunnerAssetReader assetReader;
 
@@ -57,6 +59,7 @@ main() {
         assetReader.cacheStringAsset(id, sources[node.primaryInput]);
       }
       tmpDir = await Directory.systemTemp.createTemp('build_tests');
+      anotherTmpDir = await Directory.systemTemp.createTemp('build_tests');
     });
 
     tearDown(() async {
@@ -64,8 +67,8 @@ main() {
     });
 
     test('creates a valid merged output directory', () async {
-      var success = await createMergedOutputDir(
-          tmpDir.path, graph, packageGraph, assetReader, environment, phases);
+      var success = await createMergedOutputDirectories({tmpDir.path: null},
+          graph, packageGraph, assetReader, environment, phases);
       expect(success, isTrue);
 
       _expectAllFiles(tmpDir);
@@ -76,12 +79,100 @@ main() {
           graph.get(new AssetId('b', 'lib/c.txt.copy')) as GeneratedAssetNode;
       node.isDeleted = true;
 
-      var success = await createMergedOutputDir(
-          tmpDir.path, graph, packageGraph, assetReader, environment, phases);
+      var success = await createMergedOutputDirectories({tmpDir.path: null},
+          graph, packageGraph, assetReader, environment, phases);
       expect(success, isTrue);
 
       var file = new File(p.join(tmpDir.path, 'packages/b/c.txt.copy'));
       expect(file.existsSync(), isFalse);
+    });
+
+    test('can create multiple merged directories', () async {
+      var success = await createMergedOutputDirectories(
+          {tmpDir.path: null, anotherTmpDir.path: null},
+          graph,
+          packageGraph,
+          assetReader,
+          environment,
+          phases);
+      expect(success, isTrue);
+
+      _expectAllFiles(tmpDir);
+      _expectAllFiles(anotherTmpDir);
+    });
+
+    test('removes the provided root from the output path', () async {
+      var success = await createMergedOutputDirectories({
+        tmpDir.path: 'web',
+      }, graph, packageGraph, assetReader, environment, phases);
+      expect(success, isTrue);
+
+      var webFiles = <String, dynamic>{
+        'b.txt': 'b',
+        'b.txt.copy': 'b',
+      };
+
+      _expectFiles(webFiles, tmpDir);
+    });
+
+    test('does not output the input directory', () async {
+      var success = await createMergedOutputDirectories({
+        tmpDir.path: 'web',
+      }, graph, packageGraph, assetReader, environment, phases);
+      expect(success, isTrue);
+
+      expect(new Directory(p.join(tmpDir.path, 'web')).existsSync(), isFalse);
+    });
+
+    test('outputs the packages when input root is provided', () async {
+      var success = await createMergedOutputDirectories(
+          {tmpDir.path: 'web', anotherTmpDir.path: 'foo'},
+          graph,
+          packageGraph,
+          assetReader,
+          environment,
+          phases);
+      expect(success, isTrue);
+
+      var webFiles = <String, dynamic>{
+        'packages/a/a.txt': 'a',
+        'packages/a/a.txt.copy': 'a',
+        'packages/b/c.txt': 'c',
+        'packages/b/c.txt.copy': 'c',
+        '.packages': 'a:packages/a/\r\nb:packages/b/\r\n\$sdk:packages/\$sdk/',
+      };
+
+      _expectFiles(webFiles, tmpDir);
+    });
+
+    test('only outputs files contained in the provided root', () async {
+      var success = await createMergedOutputDirectories(
+          {tmpDir.path: 'web', anotherTmpDir.path: 'foo'},
+          graph,
+          packageGraph,
+          assetReader,
+          environment,
+          phases);
+      expect(success, isTrue);
+
+      var webFiles = <String, dynamic>{
+        'b.txt': 'b',
+        'b.txt.copy': 'b',
+      };
+
+      var webNoFiles = new Set<String>()..addAll(['d.txt', 'd.txt.copy']);
+
+      var fooFiles = <String, dynamic>{
+        'd.txt': 'd',
+        'd.txt.copy': 'd',
+      };
+
+      var fooNoFiles = new Set<String>()..addAll(['b.txt', 'b.txt.copy']);
+
+      _expectFiles(webFiles, tmpDir);
+      _expectNoFiles(webNoFiles, tmpDir);
+      _expectFiles(fooFiles, anotherTmpDir);
+      _expectNoFiles(fooNoFiles, anotherTmpDir);
     });
 
     test('doesnt write files that werent output', () async {
@@ -90,8 +181,8 @@ main() {
       node.wasOutput = false;
       node.isFailure = false;
 
-      var success = await createMergedOutputDir(
-          tmpDir.path, graph, packageGraph, assetReader, environment, phases);
+      var success = await createMergedOutputDirectories({tmpDir.path: null},
+          graph, packageGraph, assetReader, environment, phases);
       expect(success, isTrue);
 
       var file = new File(p.join(tmpDir.path, 'packages/b/c.txt.copy'));
@@ -108,15 +199,15 @@ main() {
       test('fails in non-interactive mode', () async {
         environment =
             new TestBuildEnvironment(reader: assetReader, throwOnPrompt: true);
-        var success = await createMergedOutputDir(
-            tmpDir.path, graph, packageGraph, assetReader, environment, phases);
+        var success = await createMergedOutputDirectories({tmpDir.path: null},
+            graph, packageGraph, assetReader, environment, phases);
         expect(success, isFalse);
       });
 
       test('can skip creating the directory', () async {
         environment.nextPromptResponse = 0;
-        var success = await createMergedOutputDir(
-            tmpDir.path, graph, packageGraph, assetReader, environment, phases);
+        var success = await createMergedOutputDirectories({tmpDir.path: null},
+            graph, packageGraph, assetReader, environment, phases);
         expect(success, isFalse,
             reason: 'Skipping creation of the directory should be considered a '
                 'failure.');
@@ -130,8 +221,8 @@ main() {
 
       test('can delete the entire existing directory', () async {
         environment.nextPromptResponse = 1;
-        var success = await createMergedOutputDir(
-            tmpDir.path, graph, packageGraph, assetReader, environment, phases);
+        var success = await createMergedOutputDirectories({tmpDir.path: null},
+            graph, packageGraph, assetReader, environment, phases);
         expect(success, isTrue);
 
         expect(garbageFile.existsSync(), isFalse);
@@ -140,8 +231,8 @@ main() {
 
       test('can merge into the existing directory', () async {
         environment.nextPromptResponse = 2;
-        var success = await createMergedOutputDir(
-            tmpDir.path, graph, packageGraph, assetReader, environment, phases);
+        var success = await createMergedOutputDirectories({tmpDir.path: null},
+            graph, packageGraph, assetReader, environment, phases);
         expect(success, isTrue);
 
         expect(garbageFile.existsSync(), isTrue,
@@ -152,19 +243,7 @@ main() {
   });
 }
 
-void _expectAllFiles(Directory dir) {
-  var expectedFiles = <String, dynamic>{
-    'packages/a/a.txt': 'a',
-    'packages/a/a.txt.copy': 'a',
-    'packages/b/c.txt': 'c',
-    'packages/b/c.txt.copy': 'c',
-    'web/b.txt': 'b',
-    'web/b.txt.copy': 'b',
-    'web/.packages': '''
-a:packages/a/\r
-b:packages/b/\r
-\$sdk:packages/\$sdk/''',
-  };
+void _expectFiles(Map<String, dynamic> expectedFiles, Directory dir) {
   expectedFiles['.build.manifest'] =
       allOf(expectedFiles.keys.map(contains).toList());
   expectedFiles.forEach((path, content) {
@@ -173,4 +252,26 @@ b:packages/b/\r
     expect(file.readAsStringSync(), content,
         reason: 'Incorrect content for file at $path');
   });
+}
+
+void _expectNoFiles(Set<String> expectedFiles, Directory dir) {
+  for (var path in expectedFiles) {
+    var file = new File(p.join(dir.path, path));
+    expect(!file.existsSync(), isTrue, reason: 'File found at $path.');
+  }
+}
+
+void _expectAllFiles(Directory dir) {
+  var expectedFiles = <String, dynamic>{
+    'foo/d.txt': 'd',
+    'foo/d.txt.copy': 'd',
+    'packages/a/a.txt': 'a',
+    'packages/a/a.txt.copy': 'a',
+    'packages/b/c.txt': 'c',
+    'packages/b/c.txt.copy': 'c',
+    'web/b.txt': 'b',
+    'web/b.txt.copy': 'b',
+    'web/.packages': 'a:packages/a/\r\nb:packages/b/\r\n\$sdk:packages/\$sdk/',
+  };
+  _expectFiles(expectedFiles, dir);
 }